allow user to reposition boundary when creating a game Fixes #134
