+++ lib/extensions/location_extension.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/material.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 import 'package:niira/models/location.dart';
 
@@ -8,4 +9,23 @@ extension LocationExt on Location {
         target: toLatLng(),
         zoom: 17,
       );
+
+  Set<Circle> toCircles() => <Circle>{
+        Circle(
+          circleId: CircleId('boundary'),
+          center: toLatLng(),
+          radius: 100,
+          strokeWidth: 3,
+          strokeColor: Color.fromRGBO(247, 153, 0, 1),
+          fillColor: Color.fromRGBO(247, 153, 0, 0.2),
+        ),
+        Circle(
+          circleId: CircleId('player_position'),
+          center: toLatLng(),
+          radius: 8,
+          strokeWidth: 2,
+          strokeColor: Colors.white,
+          fillColor: Color.fromRGBO(130, 250, 184, 1),
+        ),
+      };
 }
